move helper from plaid_controller to finchplaid service
